pyproject.toml
[tool.poetry]
name = "stackai-auth-poc"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.31"
fastapi = "^0.111.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
sqlalchemy-mixins = "^2.0.5"
supabase = "^2.5.3"
requests = "^2.32.3"
alembic = "^1.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
